changed allocate method because of optimization of cpu usage
fixed some minor memory and string issues

+++ libs/utils/join.c
@@ -3,10 +3,10 @@
 
 #include <utils.h>
 
-size_t calculate_join(char **value, size_t size, const char *separator) {
+size_t calculate_join(char **value, unsigned short size, const char *separator) {
 	size_t source_length = (strlen(separator) * (size - 1));
 
-	for (size_t i = 0; size != 0; ++i) {
+	for (unsigned short i = 0; size != 0; ++i) {
 		const size_t string_length = strlen(value[i]);
 
 		if (string_length != 0) {
@@ -18,11 +18,11 @@ size_t calculate_join(char **value, size_t size, const char *separator) {
 	return source_length;
 }
 
-size_t join(char **value, char *source, size_t size, const char *separator) {
+size_t join(char **value, char *source, unsigned short size, const char *separator) {
 	const size_t separator_length = strlen(separator);
 	const size_t source_length = (separator_length * (size - 1));
 
-	for (size_t i = 0; size != 0; ++i) {
+	for (unsigned short i = 0; size != 0; ++i) {
 		const char *data = value[i];
 		const size_t length = strlen(data);
 		const bool has_separator = (size != 1);
